requirements.txt
Django==5.0.6
django-ckeditor==6.7.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.9.0
